requirements.txt
flask
flask-cors
redis
qdrant-client
requests
python-dotenv
uuid
google-generativeai
jinja2
google-generativeai
feedparser
certifi